fix($warning_msg): fix review

+++ myopia-business/bootstrap/src/main/resources/db/migration/V20210617__V2.0.2-vision-warning-msg.sql
@@ -6,7 +6,7 @@ CREATE TABLE `m_warning_msg` (
   `phone_numbers` varchar(255) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci DEFAULT '' COMMENT '电话号码(发送的时候才记录)',
   `send_status` tinyint NOT NULL DEFAULT '0' COMMENT '发送状态,-1发送失败,0准备发送,1是发送成功,2是取消发送',
   `send_time` timestamp NULL DEFAULT NULL COMMENT '待发送的时间',
-  `send_times` int unsigned NOT NULL COMMENT '发送次数',
+  `send_times` tinyint unsigned NOT NULL COMMENT '发送次数',
   `update_time` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP COMMENT '更新时间',
   `create_time` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT '创建时间',
   `send_day_of_year` char(7) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NOT NULL COMMENT '发送时间(yyyyD)',
